FIX: Consolidate identical variants between files
When querying a dataset, if a variant is represented in multiple vcf files
the variant count will be incremented by each file.

Count identical variants (same position, reference and alt) as 1 variant
when querying variants.

NB the variants are still calculated separately when creating dataset
summaries, as we have not yet found an efficient way to combine the variants
of entire files on the fly.