FIX: Improve duplicateVariantSearch scalability
duplicateVariantSearch stores the data for all files that it handles
simultaneously. This means the memory requirement scales linearly with
the number of input files. This means a single function can't handle
more than around 12 seconds worth of data before running our of memory.

Download and process sets of files in sequence, so only 4 are in memory
at a time. This allows processing an arbitrarily large amount data in a
single function, with the only limit being timeouts and potentially more
unique variants than can fit in memory, although this is unlikely given
the small file sizes.