FEAT: Allow summariseDataset to separate by contig
All variant files for all contigs are simultaneously listed and collated
before any calls to duplicateVariantSearch are made. In the case of a
very large dataset, this can mean that too much memory is used simply
holding all the individual variant files and their associate
vcfRegionData objects.

Allow compartmentalisation of the task by rearranging the format of the
variant files to list the contig first. This means that summariseDataset
can iterate over all the contigs, listing, collating and calling
duplicateVariantSearch for the variant files in each contig one contig
at a time. This will reduce the memory footprint of the function (for
the human genome) by about 10x.