FEAT: Increase the load on duplicateVariantSearch
A large part of the expense and time is handling the multitude of small
duplicateVariantSearch invocations. These invocations no longer need to
be small because the memory handling has changed to no longer require
memory proportional to the total of all the file sizes, but only
proportional to the single largest file (and total set of included
variants, but this is a small part).

Increase the total size of files that can be assigned to a single
invocation of duplicateVariantSearch, making the runtime become
proportionally longer to allow the number of invocations to become
proportionally smaller.