FIX: Mark zstream as complete when no more data
The gzip inflation still reports that there is data remaining when the
only data left is the final footer. This means the program tries to
extract more data and crashes when there is none forthcoming.

When the only data left to inflate is the final footer (which will be
inflated to 0 output bytes), mark the inflation as complete.