HOTFIX: comment out potentially incorrect range check
In the `perform_query` function of the `query_engine.py` file within our lambda service, we've commented out the code block that checks if a VCF position (plus reference length) falls within the defined end range. This action was taken because there is a suspicion that this block of code may be functioning incorrectly. Further investigation is required to validate its correctness and apply the necessary fix.